got rid of all the junk
