+++ app/src/main/AndroidManifest.xml
@@ -4,10 +4,8 @@
     <!-- Camara & Storage permissions to get image from Camera or Gallery -->
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.CAMERA" />
-    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
-
-    // firebase firestore
-    <uses-permission android:name="android.permission.INTERNET"/>
+    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" /> <!-- firebase firestore -->
+    <uses-permission android:name="android.permission.INTERNET" />
 
     <application
         android:allowBackup="true"
@@ -18,6 +16,15 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.GroceryManagerApp"
         tools:targetApi="31">
+        <activity
+            android:name=".SettingsActivity"
+            android:exported="false" />
+        <activity
+            android:name=".AddActivity"
+            android:exported="false" />
+        <activity
+            android:name=".ScannerPage"
+            android:exported="false" />
         <activity
             android:name=".activity_scanner_page2"
             android:exported="false">
