+++ app/src/main/java/project/stn991614740/grocerymanagerapp/AddActivity.kt
@@ -3,9 +3,9 @@ package project.stn991614740.grocerymanagerapp
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 
-class InventoryPage : AppCompatActivity() {
+class AddActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.oldactivity_scanner_page)
+        setContentView(R.layout.activity_add)
     }
 }
