--- app/src/main/java/project/stn991614740/grocerymanagerapp/DataConfirmationPage.kt
@@ -1,44 +1,0 @@
-package project.stn991614740.grocerymanagerapp
-
-import android.app.DatePickerDialog
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.widget.Button
-import android.widget.TextView
-import java.text.SimpleDateFormat
-import java.util.*
-
-class DataConfirmationPage : AppCompatActivity() {
-
-    private lateinit var tvDatePicker: TextView
-    private lateinit var btnDatePicker: Button
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.oldactivity_data_confirmation_page)
-
-        tvDatePicker = findViewById(R.id.tvDate)
-        btnDatePicker = findViewById(R.id.btnDatePicker)
-
-        val myCalendar = Calendar.getInstance()
-
-        val datePicker = DatePickerDialog.OnDateSetListener { view, year, month, dayOfMonth ->
-            myCalendar.set(Calendar.YEAR, year)
-            myCalendar.set(Calendar.MONTH, month)
-            myCalendar.set(Calendar.DAY_OF_MONTH, dayOfMonth)
-            updateLabel(myCalendar)
-        }
-
-        btnDatePicker.setOnClickListener {
-            DatePickerDialog(
-                this, datePicker, myCalendar.get(Calendar.YEAR), myCalendar.get(Calendar.MONTH),
-                myCalendar.get(Calendar.DAY_OF_MONTH)
-            ).show()
-        }
-    }
-
-    private fun updateLabel(myCalendar: Calendar) {
-        val myFormat = "dd-MM-yyyy"
-        val sdf = SimpleDateFormat(myFormat, Locale.CANADA)
-        tvDatePicker.setText(sdf.format(myCalendar.time))
-    }
-}
