+++ app/src/main/java/project/stn991614740/grocerymanagerapp/MainActivity.kt
@@ -12,50 +12,11 @@ import com.google.android.material.bottomnavigation.BottomNavigationView
 
 class MainActivity : AppCompatActivity() {
 
-
-
-
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        val button = findViewById<Button>(R.id.startButton)
-        button.setOnClickListener {
-            val Intent = Intent(this, MyFridgeFragment::class.java)
-            startActivity(Intent)
-
-            val myfridgefragment = MyFridgeFragment()
-            val scanfragment = ScanFragment()
-            val dataconfirmation = DataConfirmationFragment()
-
-            val bottomnavigationview = findViewById<BottomNavigationView>(R.id.bottomNavigationView)
-
-            replaceFragment(myfridgefragment)
-
-            bottomnavigationview.setOnItemSelectedListener {
-                when(it.itemId){
-                    R.id.myfridge -> replaceFragment(myfridgefragment)
-                    R.id.scan -> replaceFragment(scanfragment)
-                    R.id.confrim -> replaceFragment(dataconfirmation)
-                }
-                true
-            }
-        }
+        val intent = Intent(this, project.stn991614740.grocerymanagerapp.ScannerPage::class.java)
+        startActivity(intent)
     }
-    private fun replaceFragment(fragment: Fragment) {
-        if(fragment != null){
-            val transaction = supportFragmentManager.beginTransaction()
-            transaction.replace(R.id.frame_layout, fragment)
-            transaction.commit()
-    }
-
-
-
-
-    fun fakeTest() {
-        //asdf asdf
-    }
-
-}
 }
