+++ app/src/main/java/project/stn991614740/grocerymanagerapp/SettingsActivity.kt
@@ -3,9 +3,9 @@ package project.stn991614740.grocerymanagerapp
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 
-class mehrab_activity : AppCompatActivity() {
+class SettingsActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.oldactivity_mehrab)
+        setContentView(R.layout.activity_settings)
     }
 }
