--- app/src/main/res/layout/fragment_scan.xml
@@ -1,185 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:padding="10dp"
-    android:orientation="vertical"
-
-    tools:context=".ScanFragment">
-    <!-- TODO: Update blank fragment layout -->
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal">
-
-        <com.google.android.material.button.MaterialButton
-            android:id="@+id/inputImageBtn"
-            android:layout_width="match_parent"
-            android:layout_weight="1"
-            app:icon="@drawable/ic_baseline_add_a_photo_24"
-            android:text="Take Image"
-            android:layout_marginEnd="5dp"
-            android:layout_height="match_parent"/>
-
-        <com.google.android.material.button.MaterialButton
-            android:id="@+id/recognizeTextBtn"
-            android:layout_width="match_parent"
-            android:layout_weight="1"
-            app:icon="@drawable/ic_baseline_document_scanner_24"
-            android:text="Process Image"
-            android:layout_marginEnd="5dp"
-            android:layout_height="match_parent"/>
-    </LinearLayout>
-
-    <ScrollView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical">
-
-            <com.google.android.material.imageview.ShapeableImageView
-                android:id="@+id/imageIv"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:adjustViewBounds="true"
-                android:src="@drawable/ic_baseline_image_24"
-                app:strokeWidth="2dp" />
-
-            <TextView
-                style="@style/TextAppearance.MaterialComponents.Headline6"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="10dp"
-                android:text="Recognized Text: " />
-
-            <EditText
-                android:id="@+id/recognizedTextEt"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:textSize="12sp" />
-
-        </LinearLayout>
-
-    </ScrollView>
-
-    <ScrollView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:orientation="horizontal">
-            <ImageButton
-                android:id="@+id/imageButton5"
-                android:layout_width="77dp"
-                android:layout_height="69dp"
-                tools:ignore="MissingConstraints"
-                tools:layout_editor_absoluteX="25dp"
-                tools:layout_editor_absoluteY="621dp"
-                tools:srcCompat="@tools:sample/backgrounds/scenic"
-                android:layout_gravity="center"/>
-
-            <ImageButton
-                android:id="@+id/imageButton6"
-                android:layout_width="77dp"
-                android:layout_height="69dp"
-                tools:ignore="MissingConstraints"
-                tools:layout_editor_absoluteX="115dp"
-                tools:layout_editor_absoluteY="621dp"
-                tools:srcCompat="@tools:sample/backgrounds/scenic"
-                android:layout_gravity="center"/>
-
-            <ImageButton
-                android:id="@+id/imageButton7"
-                android:layout_width="77dp"
-                android:layout_height="69dp"
-                tools:ignore="MissingConstraints"
-                tools:layout_editor_absoluteX="211dp"
-                tools:layout_editor_absoluteY="621dp"
-                tools:srcCompat="@tools:sample/backgrounds/scenic"
-                android:layout_gravity="center"/>
-
-            <ImageButton
-                android:id="@+id/imageButton8"
-                android:layout_width="77dp"
-                android:layout_height="69dp"
-                android:layout_gravity="center"
-                tools:ignore="MissingConstraints"
-                tools:layout_editor_absoluteX="306dp"
-                tools:layout_editor_absoluteY="621dp"
-                tools:srcCompat="@tools:sample/backgrounds/scenic" />
-
-
-        </LinearLayout>
-
-    </ScrollView>
-
-    <ScrollView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="horizontal">
-            <TextView
-                android:id="@+id/textView16"
-                android:layout_width="77dp"
-                android:layout_height="20dp"
-                android:text="MyFridge"
-                android:textColor="@color/cardview_light_background"
-                tools:ignore="MissingConstraints"
-                tools:layout_editor_absoluteX="33dp"
-                tools:layout_editor_absoluteY="690dp" />
-
-            <TextView
-                android:id="@+id/textView17"
-                android:layout_width="77dp"
-                android:layout_height="20dp"
-                android:text="Browse"
-                android:textColor="@color/cardview_light_background"
-                tools:ignore="MissingConstraints"
-                tools:layout_editor_absoluteX="125dp"
-                tools:layout_editor_absoluteY="690dp" />
-
-            <TextView
-                android:id="@+id/textView18"
-                android:layout_width="77dp"
-                android:layout_height="20dp"
-                android:text="Scan"
-                android:textColor="@color/cardview_light_background"
-                tools:ignore="MissingConstraints"
-                tools:layout_editor_absoluteX="232dp"
-                tools:layout_editor_absoluteY="690dp" />
-
-            <TextView
-                android:id="@+id/textView19"
-                android:layout_width="77dp"
-                android:layout_height="20dp"
-                android:text="Settings"
-                android:textColor="@color/cardview_light_background"
-                tools:ignore="MissingConstraints"
-                tools:layout_editor_absoluteX="317dp"
-                tools:layout_editor_absoluteY="690dp" />
-
-        </LinearLayout>
-    </ScrollView>
-
-
-    <Button
-        android:id="@+id/button"
-        android:layout_width="328dp"
-        android:layout_height="56dp"
-        android:text="Enter Expiry Date Manually"
-        tools:ignore="MissingConstraints"
-        tools:layout_editor_absoluteX="41dp"
-        tools:layout_editor_absoluteY="516dp"
-        android:layout_gravity="center"/>
-
-</LinearLayout>
